rename: Util 패키지 이동, 공고글 목록/상세조회(피커/작성자) board 패키지로 이동
