+++ src/main/java/pickup_shuttle/pickup/domain/board/BoardService.java
@@ -25,7 +25,8 @@ import pickup_shuttle.pickup.domain.store.Store;
 import pickup_shuttle.pickup.domain.store.StoreRepository;
 import pickup_shuttle.pickup.domain.user.User;
 import pickup_shuttle.pickup.domain.user.repository.UserRepository;
-import pickup_shuttle.pickup.domain.utils.Utils;
+import pickup_shuttle.pickup.utils.Utils;
+
 
 import java.lang.reflect.Field;
 import java.time.ZoneOffset;
