+++ src/main/java/pickup_shuttle/pickup/domain/board/dto/response/ReadPickerBoardListRp.java
@@ -1,13 +1,13 @@
-package pickup_shuttle.pickup.domain.user.dto.response;
+package pickup_shuttle.pickup.domain.board.dto.response;
 
 import lombok.Builder;
 
 @Builder
-public record ReadWriterBoardListRp(
+public record  ReadPickerBoardListRp(
         Long boardId,
         String shopName,
         String destination,
-        Long finishedAt,
         int tip,
+        Long finishedAt,
         boolean isMatch
 ) { }
