+++ src/main/java/pickup_shuttle/pickup/domain/board/dto/response/ReadWriterBoard.java
@@ -1,0 +1,4 @@
+package pickup_shuttle.pickup.domain.board.dto.response;
+
+public interface ReadWriterBoard {
+}
