+++ src/main/java/pickup_shuttle/pickup/domain/user/UserController.java
@@ -11,6 +11,10 @@ import org.springframework.web.servlet.ModelAndView;
 import pickup_shuttle.pickup._core.utils.ApiUtils;
 import pickup_shuttle.pickup._core.utils.CustomPage;
 import pickup_shuttle.pickup.config.Login;
+import pickup_shuttle.pickup.domain.board.dto.response.ReadPickerBoardListRp;
+import pickup_shuttle.pickup.domain.board.dto.response.ReadPickerBoardRp;
+import pickup_shuttle.pickup.domain.board.dto.response.ReadWriterBoard;
+import pickup_shuttle.pickup.domain.board.dto.response.ReadWriterBoardListRp;
 import pickup_shuttle.pickup.domain.oauth2.CustomOauth2User;
 import pickup_shuttle.pickup.domain.user.dto.request.*;
 import pickup_shuttle.pickup.domain.user.dto.response.*;
