--- src/main/java/pickup_shuttle/pickup/domain/user/dto/response/ReadWriterBoard.java
@@ -1,4 +1,0 @@
-package pickup_shuttle.pickup.domain.user.dto.response;
-
-public interface ReadWriterBoard {
-}
