+++ src/main/java/pickup_shuttle/pickup/utils/Utils.java
@@ -1,4 +1,4 @@
-package pickup_shuttle.pickup.domain.utils;
+package pickup_shuttle.pickup.utils;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
